More type seperation
