+++ src/commands/math.ts
@@ -3,10 +3,10 @@
  ************************************************/
 
 import { MQNode } from "../services/keystroke";
-import { CursorOptions, NodeRef, MathspeakOptions, CharCmdsAny, InnerFields, JoinMethod, LatexCmdsAny, LatexCmdsSingleCharBuilder } from "../shared_types";
+import { CursorOptions, NodeRef, MathspeakOptions, LatexCmdsSingleCharBuilder } from "../shared_types";
 import { Anticursor, Cursor, MQSelection } from "../cursor"
 import { API, APIClasses, RootBlockMixin } from "../publicapi"
-import { Fragment, Ends, NodeBase, CharCmds, LatexCmds } from "../tree";
+import { Fragment, Ends, NodeBase } from "../tree";
 import { Direction, pray, L, R, noop } from "../utils"
 import { Parser, } from "../services/parser.util";
 import { domFrag } from "../domFragment";
@@ -16,6 +16,8 @@ import { Controller } from "../services/textarea";
 import { latexMathParser } from "src/services/latex";
 import { Digit, Letter } from "./math/basicSymbols";
 import { Options } from "src/options";
+import { CharCmdsAny, InnerFields, JoinMethod, LatexCmdsAny } from "src/pure_types";
+import { LatexCmds, CharCmds } from "./list";
 
 /**
  * Math tree node base class.
