+++ src/controller.ts
@@ -3,8 +3,9 @@ import { Cursor } from "./cursor"
 import { Controller } from "./services/textarea"
 import { Direction, L, R } from "./utils"
 import { KIND_OF_MQ } from "./publicapi"
-import { ControllerRoot, ControllerData, CursorOptions, HandlerName, ControllerEvent } from "./shared_types"
+import { ControllerRoot, ControllerData, CursorOptions, HandlerName } from "./shared_types"
 import { jQToDOMFragment } from "./domFragment"
+import { ControllerEvent } from "./pure_types"
 
 type TextareaKeyboardEventListeners = Partial<{
   [K in keyof HTMLElementEventMap]: (event: HTMLElementEventMap[K]) => any;
