+++ src/cursor.ts
@@ -12,7 +12,7 @@ JS environment could actually contain many instances. */
 import { Point, Fragment, Ends } from './tree';
 import { L, R, Direction, prayDirection, pray } from './utils';
 import { h } from './dom';
-import { NodeRef, CursorOptions, JoinMethod } from "./shared_types"
+import { NodeRef, CursorOptions } from "./shared_types"
 import { ControllerBase } from "./controller"
 import { DOMFragment, domFrag } from './domFragment';
 import { MQNode } from './services/keystroke';
@@ -20,6 +20,7 @@ import { Controller } from './services/textarea';
 import { U_ZERO_WIDTH_SPACE } from './unicode';
 import { getBoundingClientRect } from './browser';
 import { MathBlock } from './commands/math';
+import { JoinMethod } from './pure_types';
 
 //A fake cursor in the fake textbox that the math is rendered in.
 export class Anticursor extends Point {
