+++ src/services/latex.ts
@@ -1,5 +1,5 @@
 import { MQNode } from './keystroke';
-import { Fragment, LatexCmds } from "../tree"
+import { Fragment } from "../tree"
 import { L, R } from "../utils"
 import { Controller_keystroke } from './keystroke';
 import { MathBlock, VanillaSymbol } from '../commands/math';
@@ -9,6 +9,7 @@ import { Letter, Digit, PlusMinus } from 'src/commands/math/basicSymbols';
 import { jQToDOMFragment, domFrag } from 'src/domFragment';
 import { RootMathCommand } from 'src/commands/text';
 import { LatexCmdsSingleChar } from 'src/shared_types';
+import { LatexCmds } from 'src/commands/list';
 
 export class TempSingleCharNode extends MQNode {
   constructor(_char: string) {
